refactor-3 Pass 3C-1: extract GroupPickerUI from ParsekUI
Move group picker popup into dedicated UI/GroupPickerUI.cs class.
13 fields, 3 constants, and 8 methods moved. ParsekUI calls into
GroupPickerUI via OpenForRecording/OpenForChain/OpenForGroup/Draw/Close.
